Extract vscode webview into public matterviz/file-viewer library
Move the parse/render pipeline (parse_file_content, create_display,
JsonBrowser, PlotPanel, TrajectoryWithDos, detect + plot-utils) verbatim
from extensions/vscode/src/webview to src/lib/file-viewer and publish it
under the matterviz/file-viewer (+ /parse worker-safe subpath) exports,
so embedding hosts can reuse the file viewer without vendoring the
extension. The extension keeps thin re-export shims where its build
entry, host tsconfig, and tests need stable local paths.
