+++ extensions/vscode/tests/parse-entry.test.ts
@@ -1,10 +1,8 @@
-// The parse-only path was extracted from webview/main.ts into webview/parse.ts
-// (so embedding hosts can run it inside a Web Worker). These tests pin the
-// compatibility contract: main.ts re-exports the exact same functions, and the
-// direct parse.ts entry parses files identically without any Svelte imports.
+// The public parse-only entry stays worker-safe while the extension wrapper
+// re-exports the same functions for backward compatibility.
 import { describe, expect, test } from 'vitest'
 import * as main_entry from '../src/webview/main'
-import * as parse_entry from '../src/webview/parse'
+import * as parse_entry from '$lib/file-viewer/parse'
 
 describe(`webview parse entry points`, () => {
   test.each([`parse_file_content`, `parse_large_file_marker`, `base64_to_array_buffer`])(
@@ -34,7 +32,7 @@ describe(`webview parse entry points`, () => {
     const { existsSync, readFileSync } = await import(`node:fs`)
     const { dirname, resolve } = await import(`node:path`)
     const repo_root = resolve(import.meta.dirname, `../../..`)
-    const entry = resolve(import.meta.dirname, `../src/webview/parse.ts`)
+    const entry = resolve(repo_root, `src/lib/file-viewer/parse.ts`)
 
     const resolve_specifier = (specifier: string, from_file: string): string | null => {
       let base: string
