+++ src/lib/file-viewer/PlotPanel.svelte
