+++ src/lib/file-viewer/TrajectoryWithDos.svelte
