+++ src/lib/file-viewer/index.ts
@@ -1,0 +1,5 @@
+export * from './detect'
+export { default as JsonBrowser } from './JsonBrowser.svelte'
+export * from './main'
+export { default as PlotPanel } from './PlotPanel.svelte'
+export * from './plot-utils'
