+++ src/lib/file-viewer/plot-utils.ts
